Rename known address field + TeamOutput object
